added secondary check for position inside plot, still buggy on acad side some previous problems sorted

+++ ClassLibrary/Models/Parking/BuiltInParkingModel.cs
@@ -7,7 +7,7 @@ internal class BuiltInParkingModel
     public int TotalDisabledBigParkingSpaces { get; set; }
     public BuiltInParkingModel(Dictionary<string, string> parameters)
     {
-        FloorArea = Convert.ToInt32(parameters["п1_Пл_Общая_м2"]);
+        FloorArea = Convert.ToDouble(parameters["п1_Пл_Общая_м2"]);
         TotalParkingSpaces = Convert.ToInt32(parameters["п1_Кол_машиномест_всего_шт"]);
         TotalDisabledParkingSpaces = Convert.ToInt32(parameters["п1_Кол_машиномест_МГН_шт"]);
         TotalDisabledBigParkingSpaces = Convert.ToInt32(parameters["п1_Кол_машиномест_МГН_больших_шт"]);
